Refactor : 불필요한 파일을 하나로 합침
외부에서 사용하지 않을 컴포넌트를 index.js에 export 하는 것을 분자 컴포넌트로 묶음.
