+++ front/src/components/issue/detail/index.js
@@ -1,2 +1,2 @@
-export { default as IssueDetailBody } from './body/IssueDetailBody';
-export { default as IssueDetailHeader } from './header/IssueDetailHeader';
+export { default as IssueDetailBody } from './body';
+export { default as IssueDetailHeader } from './header';
